(lane_stitcher) inline geometry helpers / simplify to 4-stage pipeline (extract/residual/clean_lines/merge) / unify clean_lines over first+residual / drop no-op dedup, keep center_line trim // (merge_annotation) inline own geometry helpers // remove polyline_merge shared module
